test: split readiness stale-window coverage
